+++ homehttp/options_test.go
@@ -11,7 +11,7 @@ import (
 func TestClientOptionWithAppName(t *testing.T) {
 	config := &clientConfig{}
 	option := WithAppName("TestApp")
-	option.apply(config)
+	option.Apply(config)
 
 	assert.Equal(t, "TestApp", config.AppName)
 }
@@ -19,7 +19,7 @@ func TestClientOptionWithAppName(t *testing.T) {
 func TestClientOptionWithTimeout(t *testing.T) {
 	config := &clientConfig{}
 	option := WithTimeout(time.Second * 5)
-	option.apply(config)
+	option.Apply(config)
 
 	assert.Equal(t, time.Second*5, config.Timeout)
 }
@@ -28,7 +28,7 @@ func TestClientOptionWithLogger(t *testing.T) {
 	config := &clientConfig{}
 	logger := zerolog.Nop()
 	option := WithLogger(&logger)
-	option.apply(config)
+	option.Apply(config)
 
 	assert.Equal(t, &logger, config.Logger)
 }